+++ src/Main.java
@@ -1,3 +1,4 @@
+import java.util.InputMismatchException;
 import java.util.Scanner;
 
 public class Main {
@@ -9,17 +10,29 @@ public class Main {
         System.out.println("1. Enter");
         System.out.println("2. Exit");
         System.out.println("3. Reset counter");
-        System.out.println("4. Display current occupancy");
-        System.out.println("5. Set high occupancy (default is 100)");
-        System.out.println("6. View event history");
-        System.out.println("7. Help");
-        System.out.println("8. quit");
+        System.out.println("4. Manual occupancy correction");
+        System.out.println("5. Display current occupancy");
+        System.out.println("6. Set high occupancy (default is 100)");
+        System.out.println("7. View event history");
+        System.out.println("8. Help");
+        System.out.println("9. quit");
         System.out.println();
         System.out.println("Enter the command you would like to complete: ");
 
+        while (!keyboardInput.hasNextInt()) {
+            System.out.println("Must input one of the command numbers (ex. 9).");
+            System.out.println("Enter the command you would like to complete: ");
+            keyboardInput.next();
+        }
+
         int input = keyboardInput.nextInt();
 
-        while (input != 8) {
+        while (input < 0) {
+            System.out.println("Input cannot be negative. Try again: ");
+            input = keyboardInput.nextInt();
+        }
+
+        while (input != 9) {
             if (input == 1) {
                 tracker.enter();
 
@@ -30,6 +43,29 @@ public class Main {
                 tracker.resetCounter();
 
             } else if (input == 4) {
+                System.out.println("Current occupancy: " + tracker.getOccupancyCounter());
+                System.out.println("Insert new occupancy: ");
+
+                int newOccupancy = -1;
+                boolean checkOccupancy = false;
+                while (!checkOccupancy) {
+                    try {
+                        newOccupancy = keyboardInput.nextInt();
+                        if (newOccupancy < 0) {
+                            System.out.println("Occupancy cannot be negative. Insert new occupancy:");
+                        } else {
+                            checkOccupancy = true;
+                        }
+                    } catch (InputMismatchException exp) {
+                        System.out.println("Error: Input should be an integer. Try again: ");
+                        keyboardInput.next();
+                    }
+                }
+
+                tracker.manualOccupancyCorrection(newOccupancy);
+                System.out.println("Occupancy corrected to: " + tracker.getOccupancyCounter());
+
+            } else if (input == 5) {
                 System.out.println("Current occupancy: " + tracker.getOccupancyCounter());
 
                 double occupancyPercentage = tracker.getOccupancyPercentage();
@@ -51,12 +87,26 @@ public class Main {
                 else
                     System.out.println("High occupancy");
 
-            } else if (input == 5) {
-                System.out.println("Insert high occupancy: ");
-                int newHighOccupancy = keyboardInput.nextInt();
-                tracker.setHighOccupancy(newHighOccupancy);
-
             } else if (input == 6) {
+                System.out.println("Insert high occupancy: ");
+
+                boolean checkHighOccupancy = false;
+                while (!checkHighOccupancy) {
+                    try {
+                        int newHighOccupancy = keyboardInput.nextInt();
+                        if (newHighOccupancy <= 0) {
+                            System.out.println("Occupancy cannot be 0 or less. Try again.");
+                        } else {
+                            tracker.setHighOccupancy(newHighOccupancy);
+                            checkHighOccupancy = true;
+                        }
+                    } catch (InputMismatchException exp){
+                        System.out.println("Error: Input should be an integer. Try again: ");
+                        keyboardInput.next();
+                    }
+                }
+
+            } else if (input == 7) {
                 if (tracker.getEventHistory().isEmpty()) {
                     System.out.println("No events have been recorded yet.");
                 } else {
@@ -65,7 +115,7 @@ public class Main {
                     }
                 }
 
-            } else if (input == 7) {
+            } else if (input == 8) {
                 help();
 
             } else {
@@ -74,7 +124,18 @@ public class Main {
 
             System.out.println();
             System.out.println("Insert next command: ");
-            input = keyboardInput.nextInt();
+
+            boolean checkInteger = false;
+            while (!checkInteger) {
+                try {
+                    input = keyboardInput.nextInt();
+                    checkInteger = true;
+                } catch (InputMismatchException exp) {
+                    System.out.println("Input must be one of the available command numbers. Try again.");
+                    System.out.println("Insert next command: ");
+                    keyboardInput.next();
+                }
+            }
         }
 
         keyboardInput.close();
@@ -85,10 +146,11 @@ public class Main {
         System.out.println("1. Enter");
         System.out.println("2. Exit");
         System.out.println("3. Reset counter");
-        System.out.println("4. Display current occupancy");
-        System.out.println("5. Set high occupancy");
-        System.out.println("6. View event history");
-        System.out.println("7. Help");
-        System.out.println("8. quit");
+        System.out.println("4. Manual occupancy correction");
+        System.out.println("5. Display current occupancy");
+        System.out.println("6. Set high occupancy");
+        System.out.println("7. View event history");
+        System.out.println("8. Help");
+        System.out.println("9. quit");
     }
 }